+++ crates/symbolicator-native/tests/integration/source_errors.rs
@@ -9,7 +9,7 @@ use crate::{Server, example_request, setup_service};
 #[tokio::test]
 async fn test_download_errors() {
     let (symbolication, _cache_dir) = setup_service(|config| {
-        config.max_download_timeout = Duration::from_millis(200);
+        config.timeouts.max_download = Duration::from_millis(200);
     });
 
     let hitcounter = Server::new();
@@ -127,7 +127,7 @@ async fn test_deny_list() {
         config.caches.derived.retry_misses_after = Some(Duration::ZERO);
         // FIXME: `object_meta` caches treat download errors as `malformed`
         config.caches.derived.retry_malformed_after = Some(Duration::ZERO);
-        config.max_download_timeout = Duration::from_millis(200);
+        config.timeouts.max_download = Duration::from_millis(200);
         config.deny_list_time_window = Duration::from_millis(500);
         config.deny_list_bucket_size = Duration::from_millis(100);
         config.deny_list_threshold = 2;
